+++ lib/geordi/commands/bundle_install.rb
@@ -2,6 +2,6 @@ desc 'bundle-install', 'Run bundle install if required', hide: true
 def bundle_install
   if File.exist?('Gemfile') && !system('bundle check > /dev/null 2>&1')
     Interaction.announce 'Bundling'
-    Util.run! 'bundle install'
+    Util.run!('bundle install')
   end
 end