+++ lib/geordi/commands/create_databases.rb
@@ -6,11 +6,12 @@ def create_databases
   Interaction.announce 'Creating databases'
 
   if File.exist?('config/database.yml')
-    command = Util.binstub 'rake'
-    command << ' db:create:all'
-    command << ' parallel:create' if Util.file_containing?('Gemfile', /parallel_tests/)
+    command = []
+    command << Util.binstub_or_fallback('rake')
+    command << 'db:create:all'
+    command << 'parallel:create' if Util.file_containing?('Gemfile', /parallel_tests/)
 
-    Util.run! command
+    Util.run!(command)
   else
     puts 'config/database.yml does not exist. Nothing to do.'
   end