+++ lib/geordi/commands/rake.rb
@@ -14,10 +14,12 @@ def rake(*args)
 
   %w[development test cucumber].each do |env| # update long_desc when changing this
     if File.exist? "config/environments/#{env}.rb"
-      call = [Util.binstub('rake')] + args + ["RAILS_ENV=#{env}"]
-      Interaction.note_cmd call.join(' ')
+      command = []
+      command << Util.binstub_or_fallback('rake')
+      command += args
+      command << "RAILS_ENV=#{env}"
 
-      Util.run! *call
+      Util.run!(command, show_cmd: true)
     end
   end
 end