+++ lib/geordi/commands/rspec.rb
@@ -14,7 +14,7 @@ def rspec(*files)
 
     if Util.file_containing?('Gemfile', /parallel_tests/) && files.empty?
       Interaction.note 'All specs at once (using parallel_tests)'
-      Util.run! Util.binstub('rake'), 'parallel:spec', fail_message: 'Specs failed.'
+      Util.run!([Util.binstub_or_fallback('rake'), 'parallel:spec'], fail_message: 'Specs failed.')
 
     else
       # tell which specs will be run
@@ -28,13 +28,13 @@ def rspec(*files)
       command = if File.exist?('script/spec')
         ['bundle exec spec -c'] # RSpec 1
       else
-        [Util.binstub('rspec')]
+        [Util.binstub_or_fallback('rspec')]
       end
       command << '-r rspec_spinner -f RspecSpinner::Bar' if Util.file_containing?('Gemfile', /rspec_spinner/)
       command << files.join(' ')
 
       puts
-      Util.run! command.join(' '), fail_message: 'Specs failed.'
+      Util.run!(command.join(' '), fail_message: 'Specs failed.')
     end
   else
     Interaction.note 'RSpec not employed.'